refactor: change filess names to unique style
