+++ __tests__/Button.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Button from '../src/components/UI/button/Button'
+import Button from '../src/components/UI/Button/Button'
 
 test('Button renders correctly', () => {
   const tree = renderer
