+++ __tests__/Container.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Container from '../src/components/UI/container/Container';
+import Container from '../src/components/UI/Container/Container';
 
 test('Empty container renders correctly', () => {
   const tree = renderer
