+++ __tests__/ErrorMessage.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import ErrorMessage from '../src/components/UI/error-message/ErrorMessage';
+import ErrorMessage from '../src/components/UI/ErrorMessage/ErrorMessage';
 
 test('Error message renders correctly', () => {
   const tree = renderer
