+++ __tests__/Form.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Form from '../src/components/UI/forms/form/Form';
+import Form from '../src/components/UI/forms/Form/Form';
 
 test('Form renders correctly', () => {
   const tree = renderer
