+++ __tests__/H1.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import H1 from '../src/components/UI/titles/h1/H1';
+import H1 from '../src/components/UI/titles/H1/H1';
 
 test('H1 title renders correctly', () => {
   const tree = renderer
