+++ __tests__/LoginForm.test.js
@@ -1,15 +1,11 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import LoginForm from '../src/components/UI/forms/loginForm/LoginForm';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import LoginForm from '../src/components/UI/forms/LoginForm/LoginForm';
+jest.mock('react-redux');
 
 test('LoginForm renders correctly', () => {
   const tree = renderer
-    .create(<Router><Routes><Route
-      path="/login"
-      element={<LoginForm />}
-    /></Routes></Router>)
+    .create(<LoginForm />)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });
-
