+++ __tests__/Popup.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Popup from '../src/components/UI/popup/Popup';
+import Popup from '../src/components/UI/Popup/Popup';
 
 test('Popup renders correctly', () => {
   const tree = renderer
