+++ src/pages/MainPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Container from '../components/UI/container/Container';
-import H1 from '../components/UI/titles/h1/H1';
+import Container from '../components/UI/Container/Container';
+import H1 from '../components/UI/titles/H1/H1';
 
 const MainPage: React.FC = () => {
   return (
